backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0

# Web3 and Blockchain
web3==6.11.1
eth-account==0.9.0
eth-typing==3.5.1
eth-utils==2.3.0

# IPFS
ipfshttpclient==0.8.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
pydantic==2.4.2
email-validator==2.1.0.post1
python-dateutil==2.8.2
PyYAML==6.0.1

# Development Tools
black==23.10.1
flake8==6.1.0
isort==5.12.0
mypy==1.6.1